chore: move open telemetry instrumentation to separate source file
Now we will record open telemetry metrics only if explicitly turned on.
